scripts/requirements.txt
psycopg2-binary
flask
jsonschema
groq
python-dotenv

